[PP-148] Handling Multiple Fonts on Screen

+++ Project-Pentagram/Engine/Scene/MenuScene.h
@@ -11,7 +11,7 @@ class MenuScene : public GameScene
 	public:
 		virtual void GameSceneLoad() override;
 		virtual void GameSceneInit() override;
-		virtual void GameSceneUpdate(double dt) override;
+		virtual void GameSceneUpdate(float dt) override;
 		virtual void GameSceneDraw() override;
 		virtual void GameSceneUnload() override;
 		virtual void GameSceneFree() override;
