+++ Project-Pentagram/Engine/Scene/NodeScene.h
@@ -9,7 +9,7 @@ class NodeScene : public GameScene
 	public:
 		virtual void GameSceneLoad() override;
 		virtual void GameSceneInit() override;
-		virtual void GameSceneUpdate(double dt) override;
+		virtual void GameSceneUpdate(float dt) override;
 		virtual void GameSceneDraw() override;
 		virtual void GameSceneUnload() override;
 		virtual void GameSceneFree() override;
